add: merge leo implementation as openmp_mem_efficient and refactor benchmark, mnist benchmark code
