+++ all_benchmark.sh
@@ -1,0 +1,2 @@
+bash benchmark.sh
+bash mnist_benchmark.sh
