--- pytorch_extension/include/config.h
@@ -1,4 +1,0 @@
-#pragma once
-
-// if verbose is set to 1, then the program will print out the time information
-#define VERBOSE 0
