+++ pytorch_extension/sparse_mm.cpp
@@ -1,13 +1,19 @@
 #include <torch/extension.h>
 #include <iostream>
-#include <thread>
 #include <builtin.h>
-#include <builtin_parallel_structure.h>
-#include <builtin_omp.h>
-#include <builtin_std_thread.h>
+#include <parallel_structure.h>
+#include <omp_mm.h>
+#include <omp_mem_efficient.h>
+#include <std_thread.h>
 #include <logger.h>
 
-const unsigned int THREADS = std::thread::hardware_concurrency();
+// if you want to show the logs, you can set VERBOSE to 1
+int VERBOSE = 0;
+
+void set_verbose(int verbose)
+{
+    VERBOSE = verbose;
+}
 
 torch::Tensor sparse_mm(torch::Tensor sparse_matrix_0, torch::Tensor sparse_matrix_1, TestConfig config)
 {
@@ -15,7 +21,7 @@ torch::Tensor sparse_mm(torch::Tensor sparse_matrix_0, torch::Tensor sparse_matr
     TORCH_CHECK(sparse_matrix_0.is_sparse(), "sparse_matrix_0 must be a sparse tensor");
     TORCH_CHECK(sparse_matrix_1.is_sparse(), "sparse_matrix_1 must be a sparse tensor");
 
-    torch::Tensor answer = sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1);
+    torch::Tensor answer = builtin::sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1);
     // logger.showLogs(config);
     // because it is hard to compare with the parallel structure code, so we compare the builtin implementation by using parallel structure refactoring code
     return answer;
@@ -27,8 +33,9 @@ torch::Tensor parallel_structure_sparse_mm(torch::Tensor sparse_matrix_0, torch:
     TORCH_CHECK(sparse_matrix_0.is_sparse(), "sparse_matrix_0 must be a sparse tensor");
     TORCH_CHECK(sparse_matrix_1.is_sparse(), "sparse_matrix_1 must be a sparse tensor");
 
-    torch::Tensor answer = sparse_sparse_matmul_cpu_parallel_structure(sparse_matrix_0, sparse_matrix_1);
-    logger.showLogs(config);
+    torch::Tensor answer = parallel_structure::sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1);
+    if (VERBOSE)
+        logger.showLogs(config);
     return answer;
 }
 
@@ -38,9 +45,23 @@ torch::Tensor openmp_sparse_mm(torch::Tensor sparse_matrix_0, torch::Tensor spar
     TORCH_CHECK(sparse_matrix_0.is_sparse(), "sparse_matrix_0 must be a sparse tensor");
     TORCH_CHECK(sparse_matrix_1.is_sparse(), "sparse_matrix_1 must be a sparse tensor");
 
-    torch::Tensor answer = sparse_sparse_matmul_cpu_omp(sparse_matrix_0, sparse_matrix_1, num_threads);
+    torch::Tensor answer = omp::sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1, num_threads);
 
-    logger.showLogs(config);
+    if (VERBOSE)
+        logger.showLogs(config);
+    return answer;
+}
+
+torch::Tensor openmp_mem_efficient_sparse_mm(torch::Tensor sparse_matrix_0, torch::Tensor sparse_matrix_1, const int num_threads, TestConfig config)
+{
+    logger.reset();
+    TORCH_CHECK(sparse_matrix_0.is_sparse(), "sparse_matrix_0 must be a sparse tensor");
+    TORCH_CHECK(sparse_matrix_1.is_sparse(), "sparse_matrix_1 must be a sparse tensor");
+
+    torch::Tensor answer = omp_mem_efficient::sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1, num_threads);
+
+    // different from the parallel structure code, we don't need to show the logs here
+    // logger.showLogs(config);
     return answer;
 }
 
@@ -50,8 +71,10 @@ torch::Tensor std_thread_sparse_mm(torch::Tensor sparse_matrix_0, torch::Tensor 
     TORCH_CHECK(sparse_matrix_0.is_sparse(), "sparse_matrix_0 must be a sparse tensor");
     TORCH_CHECK(sparse_matrix_1.is_sparse(), "sparse_matrix_1 must be a sparse tensor");
 
-    torch::Tensor answer = sparse_sparse_matmul_cpu_std_thread(sparse_matrix_0, sparse_matrix_1, num_threads);
-    logger.showLogs(config);
+    torch::Tensor answer = std_thread::sparse_sparse_matmul_cpu(sparse_matrix_0, sparse_matrix_1, num_threads);
+    
+    if (VERBOSE)
+        logger.showLogs(config);
     return answer;
 }
 
@@ -69,6 +92,7 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m)
         .value("BUILTIN", TestType::BUILTIN)
         .value("PARALLEL_STRUCTURE", TestType::PARALLEL_STRUCTURE)
         .value("OPENMP", TestType::OPENMP)
+        .value("OPENMP_MEM_EFFI", TestType::OPENMP_MEM_EFFI)
         .value("STD_THREAD", TestType::STD_THREAD)
         .export_values();
 
@@ -85,5 +109,7 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m)
     m.def("sparse_mm", &sparse_mm, "Sparse matrix multiplication");
     m.def("parallel_structure_sparse_mm", &parallel_structure_sparse_mm, "Parallel Structure Refactorization Sparse matrix multiplication");
     m.def("openmp_sparse_mm", &openmp_sparse_mm, "OpenMP Sparse matrix multiplication");
+    m.def("openmp_mem_efficient_sparse_mm", &openmp_mem_efficient_sparse_mm, "OpenMP Mem Efficient Sparse matrix multiplication");
     m.def("std_thread_sparse_mm", &std_thread_sparse_mm, "std::thread Sparse matrix multiplication");
+    m.def("set_verbose", &set_verbose, "set verbose");
 }
