+++ requirements.txt
@@ -1,4 +1,5 @@
 torch
+torchvision
 pytest
 tqdm
 matplotlib
