+++ src/benchmark_utils.py
@@ -6,9 +6,9 @@ from .extension import ExtensionHandler
 
 def benchmark(func, *args):
     """Benchmark the execution time of a function"""
-    start_time = time.time()
+    start_time = time.perf_counter()
     func(*args)
-    end_time = time.time()
+    end_time = time.perf_counter()
     return end_time - start_time
 
 def builtin_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor):
@@ -26,6 +26,10 @@ def parallel_structure_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: t
 def openmp_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor, num_threads: int, test_configuration: SparseMatrixTestConfiguration):
     """Computes the product of two sparse matrices by using the openmp version of ``torch.sparse.mm`` function"""
     return ExtensionHandler.openmp_sparse_mm(sparse_matrix, sparse_matrix_1, test_configuration, num_threads)
+
+def openmp_mem_effi_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor, num_threads: int, test_configuration: SparseMatrixTestConfiguration):
+    """Computes the product of two sparse matrices by using the openmp version of ``torch.sparse.mm`` function"""
+    return ExtensionHandler.openmp_mem_effi_sparse_mm(sparse_matrix, sparse_matrix_1, test_configuration, num_threads)
 
 def std_thread_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor, num_threads: int, test_configuration: SparseMatrixTestConfiguration):
     """Computes the product of two sparse matrices by using the pthread version of ``torch.sparse.mm`` function"""
@@ -46,6 +50,7 @@ class BenchmarkResult:
     sparse_mm_extension: float
     parallel_structure_sparse_mm: float
     openmp_sparse_mm: float
+    openmp_mem_effi_sparse_mm: float
     std_thread_sparse_mm: float
     dense_mm: float
     A_col: int
@@ -64,6 +69,7 @@ class BenchmarkResult:
         return_string += f'builtin_sparse_mm_extension: {self.sparse_mm_extension*1000:>15.5f}ms\n'
         return_string += f'parallel_structure_sparse_mm: {self.parallel_structure_sparse_mm*1000:>14.5f}ms\n'
         return_string += f'openmp_sparse_mm: {self.openmp_sparse_mm*1000:>26.5f}ms\n'
+        return_string += f'openmp_mem_effi_sparse_mm: {self.openmp_mem_effi_sparse_mm*1000:>17.5f}ms\n'
         return_string += f'std_thread_sparse_mm: {self.std_thread_sparse_mm*1000:>22.5f}ms\n'
         return_string += f'dense_mm: {self.dense_mm*1000:>34.5f}ms\n'
         return_string += "-" * 50 + "\n"
