+++ src/utils.py
@@ -1,12 +1,14 @@
-import torch
-import dataclasses
 from enum import Enum
+import dataclasses
+import torch
+import time
 
 class TestType(Enum):
     BUILTIN = 0
     PARALLEL_STRUCTURE = 1
     OPENMP = 2
-    STD_THREAD = 3
+    OPENMP_MEM_EFFI = 3
+    STD_THREAD = 4
 
 @dataclasses.dataclass
 class SparseMatrixTestConfiguration(object):
@@ -49,3 +51,62 @@ def assert_sparse_or_dense_matrix_are_equal(matrix: torch.Tensor, matrix_1: torc
         matrix_1 = matrix_1.to_dense()
     assert matrix.allclose(matrix_1)
     return True
+
+class Logger:
+    def __init__(self):
+        self.start_time = dict()
+        self.end_time = dict()
+    
+    def reset(self):
+        """
+        Reset the start and end time dictionaries.
+        """
+        self.start_time.clear()
+        self.end_time.clear()
+    
+    def start(self, name):
+        """
+        Start the timer for a given name.
+
+        Parameters:
+        name (str): The name of the timer.
+        """
+        self.start_time[name] = time.perf_counter()
+    
+    def end(self, name):
+        """
+        End the timer for a given name.
+
+        Parameters:
+        name (str): The name of the timer.
+        """
+        self.end_time[name] = time.perf_counter()
+
+    def get_total(self):
+        """
+        Calculate the total time elapsed for all timers.
+
+        Returns:
+        float: The total time elapsed in seconds.
+        """
+        total = 0
+        for name in self.start_time:
+            total += self.end_time[name] - self.start_time[name]
+        return total
+
+    def show(self):
+        """
+        Print the total time elapsed for all timers in milliseconds.
+        """
+        total = 0
+        for name in self.start_time:
+            total += self.end_time[name] - self.start_time[name]
+        print(f'Total: {total*1000:.2f}ms')
+
+ConfigNumberType = int | float
+def create_config_list_by_mul_step(start: ConfigNumberType, end: ConfigNumberType, step: ConfigNumberType) -> list[ConfigNumberType]:
+    config_list = []
+    while start <= end:
+        config_list.append(start)
+        start *= step
+    return config_list
